PDF util
